+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 ### Changed
 - Upgraded `org.jetbrains.kotlin.jvm` to `1.4.31`
 - Deprecated support for IU-2020.2
+- Improved logging for CFLint configuration load and file scanning failures
 
 ## [0.2.1]
 ### Added